refactor: enhance create plan button appearance cy-89
