+++ apps/frontend/src/libs/components/dashboard/components/greeting/greeting.tsx
@@ -3,7 +3,7 @@ import { type FC } from "react";
 import { CreateDoc } from "~/assets/img/icons/icons.js";
 import { useAppSelector } from "~/libs/hooks/hooks.js";
 
-import { CreatePlanButton } from "../create-button/create-button.js";
+import { CreatePlanButton } from "../create-plan-button/create-plan-button.js";
 import styles from "./styles.module.css";
 
 const Greeting: FC = () => {
